pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.25",
    "paddleocr>=3.3.2",
    "paddlepaddle>=3.2.2",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.1",
]
